+++ RSS-CSS-Selectors/src/classes/view/main/checker.ts
@@ -2,7 +2,7 @@ import { TableView } from "./table/table-view";
 import { CssView } from "./css-view/css-view";
 import { HTMLView } from "./html-view/html-view";
 import { LevelsView } from "./levels-view/level-view";
-import { LEVELS_DATA } from "../../../data/lelels-data";
+import { LEVELS_DATA } from "../../../data/levels-data";
 import { state } from "../../state/state";
 
 const CSS_CLASSES = {
